wxSmith: Changed int to size_t in wxsCoder to prevent compiler warnings (suggested by MortenMacFly)
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@4013 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/contrib/wxSmith/wxscoder.cpp
@@ -168,7 +168,7 @@ bool wxsCoder::ApplyChanges(const wxString& FileName,const wxString& Header,cons
 
     // Detecting EOL in this sources
     wxString EOL;
-    for ( int i=0; i<Content.Length(); i++ )
+    for ( size_t i=0; i<Content.Length(); i++ )
     {
         wxChar ch = Content.GetChar(i);
         if ( ch==_T('\n') || ch==_T('\r') )
